@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 33
     defaultConfig {
         applicationId "gstavrinos.destinationalarm"
         minSdkVersion 24
@@ -25,13 +25,14 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    //noinspection GradleCompatible
+    implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'org.osmdroid:osmdroid-android:6.0.2'
-    implementation 'com.android.support:multidex:1.0.3'
-    api ("com.android.support:support-v4:27.1.1"){force=true}
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'androidx.multidex:multidex:2.0.1'
+    api ('androidx.legacy:legacy-support-v4:1.0.0'){force=true}
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
     implementation 'com.github.tbruyelle:rxpermissions:0.10.1'
     implementation 'com.jakewharton.rxbinding2:rxbinding:2.1.1'
     implementation ('com.github.MKergall:osmbonuspack:6.5.2'){
